+++ docs/com/invisiblecollector/IcApiFacade.html
@@ -134,7 +134,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr class="altColor">
 <td class="colOne"><code><span class="memberNameLink"><a href="../../com/invisiblecollector/IcApiFacade.html#IcApiFacade-com.invisiblecollector.connection.ApiRequestFacade-com.invisiblecollector.model.json.JsonModelFacade-">IcApiFacade</a></span>(com.invisiblecollector.connection.ApiRequestFacade&nbsp;apiFacade,
-           com.invisiblecollector.model.json.JsonModelFacade&nbsp;jsonFacade)</code>
+           com.invisiblecollector.model.serialization.JsonModelFacade&nbsp;jsonFacade)</code>
 <div class="block">Creates an IcApiFacade with injection</div>
 </td>
 </tr>
@@ -292,7 +292,7 @@ extends java.lang.Object</pre>
 <li class="blockList">
 <h4>IcApiFacade</h4>
 <pre>public&nbsp;IcApiFacade(com.invisiblecollector.connection.ApiRequestFacade&nbsp;apiFacade,
-                   com.invisiblecollector.model.json.JsonModelFacade&nbsp;jsonFacade)</pre>
+                   com.invisiblecollector.model.serialization.JsonModelFacade&nbsp;jsonFacade)</pre>
 <div class="block">Creates an IcApiFacade with injection
 
  <p>Preferably use <a href="../../com/invisiblecollector/IcApiFacade.html#IcApiFacade-java.lang.String-"><code>IcApiFacade(String)</code></a> or <a href="../../com/invisiblecollector/IcApiFacade.html#IcApiFacade-java.lang.String-java.net.URI-"><code>IcApiFacade(String, URI)</code></a></div>
