+++ src/test/java/com/invisiblecollector/model/serialization/JsonModelFacadeTest.java
@@ -1,4 +1,4 @@
-package com.invisiblecollector.model.json;
+package com.invisiblecollector.model.serialization;
 
 import com.invisiblecollector.exceptions.IcException;
 import com.invisiblecollector.model.Company;
